feat: Update Issue Label Generator to Issue Analysis Generator
Changelog:
- Changed class name to `IssueAnalysisGenerator`
- Add new dataclass for issue description
- Update prompt for issue label generation
- Add prompt for issue description generation
