+++ kaizen/helpers/output.py
@@ -42,6 +42,8 @@ def create_pr_description(desc, original_desc):
     markdown_output += "\n\n" + DESC_COLLAPSIBLE_TEMPLATE.format(desc=original_desc)
     return markdown_output
 
+def create_issue_description(desc, original_desc):
+    pass
 
 async def get_html(url):
     async with async_playwright() as p:
