+++ kaizen/llms/prompts/issue_analysis_prompts.py
@@ -1,0 +1,53 @@
+ISSUE_LABEL_SYSTEM_PROMPT = """
+As a senior project manager and a senior software developer reviewing raised GitHub issues, feature requests, and bug reports, make decisions by reviewing the issues and give constructive feedback and suggestions for improvement. Understand the nature of the issue and the problem being reported. Use your expertise to provide comprehensive feedback without asking clarifying questions.
+"""
+
+ISSUE_LABEL_PROMPT = """
+Understand the type of issue and choose labels from a given list to be added to the issue. The existing labels will be provided as a string in the following format:
+
+<label>: (Optional)<description>
+
+Output Format:
+```markdown
+- {{label_1}}
+- {{label_2}}
+...
+- {{label_n}}
+```
+
+Instructions:
+- Understand the labels from the list and their description.
+- Choose labels only from the provided list that best fit the issue.
+
+Based on:
+Title: {ISSUE_TITLE}
+Description: {ISSUE_DESCRIPTION}
+Exisiting labels: {ISSUE_LABEL_LIST}
+"""
+
+ISSUE_DESC_PROMPT="""
+Understand the issue and summarize the main purpose, scope of changes, modifications and approach required to tackle the issue. Also provide a brief summary of why the change proposed in the issue is necessary.
+
+Output Format:
+```markdown
+# {{Generated PR Title}}
+
+## Overview
+{{Brief summary of the issue}}
+
+## Suggestions
+- Primary actions: {{List the most critical steps to address the issue}}
+- Potential enhancements: {{List additional improvements or optimizations}}
+- Risk mitigation: {{List any risks and how to mitigate them}}
+```
+
+Instructions:
+- Create a concise summary of the issue's main purpose.
+- Use markdown formatting for readability.
+- Focus of significant suggestions and avoid repetition.
+
+Based on:
+Title: {ISSUE_TITLE}
+Description: {ISSUE_DESCRIPTION}
+"""
+
