--- kaizen/llms/prompts/issue_label_prompts.py
@@ -1,18 +1,0 @@
-ISSUE_LABEL_SYSTEM_PROMPT = """
-As a senior project manager reviewing raised GitHub issues, feature requests, and bug reports, make decisions by reviewing the issues and recommend GitHub labels to be added to those issues. Understand the nature of the issue and the problem being reported. Use your expertise to correctly choose the labels from an existing list of labels.
-"""
-
-ISSUE_LABEL_PROMPT = """
-Understand the type of issue and choose labels from a given list to be added to the issue. The existing labels will be provided as a string in the following format:
-
-<label>: (Optional)<description>
-
-Instructions:
-- Understand the labels from the list and their description.
-- Choose labels only from the provided list that best fit the issue.
-
-Based on:
-Title: {ISSUE_TITLE}
-Description: {ISSUE_DESCRIPTION}
-Exisiting labels: {ISSUE_LABEL_LIST}
-"""
